Score calculation and keywise filtering seperated, refactored.
1. Score calculation is seperated from keywise filtering through the interface IUriScoreCalculator and it's implementation class UriScoreCalculator
2. Tests also seperated accordingly.
3. Some errors induced by earlier commits are corrected.
4. Further refactoring done.
