+++ squirrel.frontier/src/main/java/org/dice_research/squirrel/queue/scorecalculator/UriScoreCalculator.java
@@ -1,4 +1,4 @@
-package org.dice_research.squirrel.queue.scorebasedfilter;
+package org.dice_research.squirrel.queue.scorecalculator;
 
 import org.aksw.jena_sparql_api.core.QueryExecutionFactory;
 import org.aksw.jena_sparql_api.http.QueryExecutionFactoryHttp;
@@ -14,27 +14,27 @@ import org.apache.jena.query.QuerySolution;
 import org.apache.jena.query.ResultSet;
 import org.apache.jena.sparql.core.DatasetDescription;
 import org.dice_research.squirrel.data.uri.CrawleableUri;
+import org.dice_research.squirrel.queue.scorebasedfilter.UriKeywiseFilter;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.net.URI;
-import java.util.*;
 
 /**
- * This class filters the {@link CrawleableUri}s to be added to the queue based on the score.
+ * Interface for getting duplicity score of a Uri.
  * The duplicity score is calculated for a {@link CrawleableUri} as 1 / (number of times the Uri occurs as a subject in graphs
  */
-public class URIKeywiseFilter implements IURIKeywiseFilter {
+public class UriScoreCalculator implements IUriScoreCalculator {
 
-    private static final Logger LOGGER = LoggerFactory.getLogger(URIKeywiseFilter.class);
+    private static final Logger LOGGER = LoggerFactory.getLogger(UriKeywiseFilter.class);
 
     protected QueryExecutionFactory queryExecFactory = null;
 
-    public URIKeywiseFilter(QueryExecutionFactory qe) {
+    public UriScoreCalculator(QueryExecutionFactory qe) {
         this.queryExecFactory = qe;
     }
 
-    public URIKeywiseFilter(String sparqlEndpointUrl, String username, String password) {
+    public UriScoreCalculator(String sparqlEndpointUrl, String username, String password) {
         if (username != null && password != null) {
             // Create the factory with the credentials
             final Credentials credentials = new UsernamePasswordCredentials(username, password);
@@ -84,27 +84,6 @@ public class URIKeywiseFilter implements IURIKeywiseFilter {
             return 1;
         }
         return 1 / (float) uriScore;
-    }
-
-    @Override
-    public Map<CrawleableUri, Float> filterUrisKeywise(Map keyWiseUris, int minNumberOfUrisToCheck, float criticalScore) {
-        Collection<List<CrawleableUri>> uriLists = keyWiseUris.values();
-        Map<CrawleableUri, Float> filteredUriMap = new HashMap<>();
-        for (List<CrawleableUri> uriList : uriLists) {
-            boolean scoresBelowCritical = true;
-            for (int i = 0; i < (minNumberOfUrisToCheck < uriList.size() ? minNumberOfUrisToCheck : uriList.size()); i++) {
-                float score = getURIScore(uriList.get(i));
-                if (score > criticalScore) {
-                    scoresBelowCritical = false;
-                }
-            }
-            if (!scoresBelowCritical) {
-                for (CrawleableUri uri : uriList) {
-                    filteredUriMap.put(uri, getURIScore(uri));
-                }
-            }
-        }
-        return filteredUriMap;
     }
 
     /**
